Allow multiple Alpine.morph calls one after the other
